requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9; platform_system!='Windows'
psycopg2-binary==2.9.9; platform_system=='Windows' and python_version >= '3.8' and python_version < '3.11'
bcrypt==4.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
Flask-WTF==1.2.1
email-validator==2.1.0
